% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCSV.R
\name{generate_readCSV}
\alias{generate_readCSV}
\title{Generate readCSV Routine}
\usage{
generate_readCSV(colClasses, module = Module())
}
\arguments{
\item{colClasses}{column classes of CSV file}

\item{module}{LLVM module}
}
\description{
Generate a routine in LLVM IR for reading a CSV file.
}

