% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCSV.R
\name{include_rdefines}
\alias{include_rdefines}
\title{Include Rdefines.h}
\usage{
include_rdefines(module = Module())
}
\arguments{
\item{module}{LLVM module}
}
\description{
Declare structures and routines from Rdefines.h and Rinternals.h.
}

