% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCSV.R
\name{include_stdio}
\alias{include_stdio}
\title{Include stdio.h}
\usage{
include_stdio(module = Module())
}
\arguments{
\item{module}{LLVM module}
}
\description{
Declare structures and routines from stdio.h.
}

