% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCSV.R
\name{include_stdlib}
\alias{include_stdlib}
\title{Include stdlib.h}
\usage{
include_stdlib(module = Module())
}
\arguments{
\item{module}{LLVM module}
}
\description{
Declare structures and routines from stdlib.h.
}

